# packages/acquirer-backend/docs/openapi-autogenerated.yaml
openapi: 3.0.0
info:
  title: Merchant Acquirer API
  version: 0.1.0
  description: Merchant Acquirer Backend
servers:
  - url: /api/v1
components:
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT
      value: Bearer <JWT token here>
paths:
...
  /config/trace-level:
    put:
      tags:
        - Server Configuration
      security:
        - Authorization: []
      summary: Set the log level
      description: This endpoint allows you to set the logging level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                level:
                  type: string
                  description: The registration status of the merchant
                  enum:
                    - error
                    - warn
                    - info
                    - http
                    - verbose
                    - debug
                    - silly
                  example: error
      responses:
        '200':
          description: Log level set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid log level
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Invalid log level: error'
...
  /users/login:
    post:
      tags:
        - Portal Users
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: test1@email.com
                  description: The email for login
                password:
                  type: string
                  example: password
                  description: The password for login in clear text
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Invalid credentials
        '422':
          description: Validation error
...